packages/transports/voice/pyproject.toml
[project]
name = "transport-voice"
version = "0.1.0"
description = "RAW Platform voice transport - Pipecat adapter for real-time voice conversations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "raw-core",
    "raw-bot",
    "pipecat-ai[deepgram,elevenlabs,local,silero]",
    "pydantic>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/transport_voice"]
